refactor: 694 - move files to ui folder
