+++ src/widgets/general/ui/general.scss
